Refact of Menu Component
